requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose==3.3.0
python-dotenv==1.0.1
pydantic[email]==2.4.2
httpx==0.26.0
python-multipart==0.0.6
SQLAlchemy==2.0.27
# Add if you use PostgreSQL:
psycopg2-binary==2.9.9
PyJWT==2.8.0 